Split target controller with emulated controller
Target is a physical one, emulated is emulated one. Don't mix them. This fixes
an issue where the code tried to unplug emulated controller on shutdown when no
physical was connected causing a crash. It also prevents any operations on a
physical one to happen on emulated one.

Fix all the code so the targetController can be null.