The (Actions != prevActions) comparison will always be true
Those are different references (they are cloned), so it will always be true. If
we didn't clone it on the other hand it would always return false and cause
issues. This would work only if proper deep comparison operator== would get
implemented for actions. And even then this callback is now only called when
something has really changed so any comparison inside is useless. Remove.