Fix an extremely rare crash on Gordon
It's related to mouse action logic and the fact that Gordon sends only one X/Y
set for both left pad and stick. With a very specific order of clicks and
touches prevValue could've been bool while it should've been already a
vector. Just don't reuse prevValue as 2 different types. Return to having
another variable for this. Actually MouseActions should be split into cursor
and clicks (like AxisActions and ButtonsActions) but that's for another day.