Proper calculation of round touchpad to dpad for Gordon
Gordon's touchpad's X/Y coordinates are canted towards the middle by 15
degrees. Additionally the touchpads are round so the math for Steam
Deck (above/below, to the left/to the right) translates poorly for the angled
directions (n * 45). Do this by the radius and angle, compensate for the 15
degree rotation (the carved dpad on left touchpads is not rotated). Also
compenstate on the right pad although not sure this should stay. Right touchpad
doesn't have carved dpad so the 15 degrees might make sense in this case. Will
see. Right pad to directions won't be widely used anyway. Left pad though is
much more usable now.