Don't touch RPAD_MARGIN, why is everybody turning it off?!
I also turned it off initially as both linux kernel and SCC as well as seemingly
dumped packets from Steam application turn it off. I thought this is outside
margin of the pad. Turns out this is micro movements filtering on the hardware
level. Turning it off you get raw sensor information. But this raw data is so
noisy it's almost impossible to filter it out without causing latency. Let the
hardware do its thing.